+++ test/fixture/define/exports-default/input.js
@@ -1,0 +1,3 @@
+define('foo', ['exports', 'bar', 'baz'], function (exports, bar, baz) {
+  exports.default = bar(baz)
+})