+++ test/fixture/define/exports-default/output.js
@@ -1,0 +1,3 @@
+import bar from 'bar';
+import baz from 'baz';
+export default bar(baz);