+++ test/fixture/define/object-define-property/input.js
@@ -1,0 +1,6 @@
+define(['exports'], function (exports) {
+  Object.defineProperty(exports, '__esModule', {
+    value: true
+  });
+  exports.default = 'foo';
+});